Undo initialisation of fields. Add first of logging code.
The previous clean-up effort included an attempt to ensure more (not
yet all) fields are set in the constructor, so code-inspector complained
less. This had side-effect of wiping out the side-effect of setUser().
Reverted the change. Should be reconsidered later.

Added some logging. This library is hard to use it a script with print
statements. Started the first use of the logging library.